fix: コードレビュー指摘 23件を修正（セキュリティ・品質・並行性）
【高優先度】
- DecompressViewModel: @MainActor 追加、progress を Task{@MainActor} でラップ
- CompressViewModel / DecompressViewModel / SendViewModel:
  startAccessingSecurityScopedResource() / defer stopAccessing() 追加
- LibArchiveWrapper: ZIP Slip 対策（パスサニタイズ・destination 配下チェック）
- GmailService: resumed フラグで continuation 二重 resume 防止
- SendViewModel: CancellationError 時に isCountingDown / countdown をリセット
- GmailAPIClient: buildMIMEMessage を async throws 化・Task.detached で非同期読み込み

【中優先度】
- AutoDeleteService: deinit 削除・UserDefaults キーを UDKey 定数に統一
- HistoryService: NSEntityDescription.insertNewObject 推奨パターンに変更
- SendViewModel: UserDefaults キーを SettingsViewModel.UDKey に統一
- PasswordGeneratorSheet: onAppear で UserDefaults 設定値を引き継ぎ
- LibArchiveWrapper: TAR 圧縮・解凍に時間ベース進捗シミュレーション追加
- GmailAPIClient: sendEndpoint をインスタンスプロパティ化（テスト容易性向上）
- CompressViewModel: 圧縮完了後に password をクリア
- HistoryView: List + ForEach.onDelete でスワイプ削除 UI 追加
- GmailService: passwordEmailDelayNanoseconds 定数化

【低優先度】
- CryptoService: generateKeyPair を CryptoKitWrapper に委譲
- KeychainService: OAuthトークン保管コメント修正
- String+Extensions: isValidEmail の IDN 非対応コメント追記
- CompressionService: errorDescription を NSLocalizedString 化
- PasswordService: acceptLimit を Int 型に統一
- CompressView: openSavePanel の冗長な @MainActor 削除
- DecompressView: プレースホルダーを if let 分岐で多言語対応
- GmailAPIClient: retryStatus の冗長な分岐を削除

+++ SecureZip/SecureZip/Services/PasswordService.swift
@@ -45,7 +45,7 @@ final class PasswordService: PasswordServiceProtocol {
 
         // rejection sampling: モジュロバイアスを除去するため
         // 256 が charsetCount で割り切れない場合、末尾の余り分のバイトを棄却して再抽選する
-        let acceptLimit = UInt8((256 / charsetCount) * charsetCount)
+        let acceptLimit = (256 / charsetCount) * charsetCount
 
         var result: [Character] = []
         result.reserveCapacity(length)
@@ -59,7 +59,7 @@ final class PasswordService: PasswordServiceProtocol {
                 continue
             }
             // acceptLimit 未満のバイトのみ採用（均等分布を保証）
-            if byte < acceptLimit {
+            if Int(byte) < acceptLimit {
                 result.append(charsetArray[Int(byte) % charsetCount])
             }
         }
